pkg/operator/encryption/kms: cleanup
