+++ pkg/operator/encryption/kms/helpers.go
@@ -2,37 +2,11 @@ package kms
 
 import (
 	"fmt"
-	"strconv"
-	"strings"
 
 	"github.com/openshift/api/features"
 	"github.com/openshift/library-go/pkg/operator/configobserver/featuregates"
 	corev1 "k8s.io/api/core/v1"
 )
-
-const pluginConfigDataKeyPrefix = "kms-plugin-config-"
-
-// ToPluginConfigSecretDataKeyFor constructs the data key for storing a KMS plugin config in the encryption-config Secret.
-// The keyID must be a valid non-negative integer string.
-func ToPluginConfigSecretDataKeyFor(keyID string) (string, error) {
-	if _, err := strconv.ParseUint(keyID, 10, 64); err != nil {
-		return "", fmt.Errorf("invalid keyID %q: must be a non-negative integer", keyID)
-	}
-	return pluginConfigDataKeyPrefix + keyID, nil
-}
-
-// KeyIDFromPluginConfigSecretDataKey extracts the keyID from a kms-plugin-config data key.
-// Returns the keyID and true if the key matches the "kms-plugin-config-<keyID>" pattern.
-func KeyIDFromPluginConfigSecretDataKey(dataKey string) (string, bool, error) {
-	keyID, found := strings.CutPrefix(dataKey, pluginConfigDataKeyPrefix)
-	if !found || len(keyID) == 0 {
-		return "", false, nil
-	}
-	if _, err := strconv.ParseUint(keyID, 10, 64); err != nil {
-		return "", false, fmt.Errorf("invalid keyID %q: must be a non-negative integer", keyID)
-	}
-	return keyID, true, nil
-}
 
 // AddKMSPluginVolumeAndMountToPodSpec conditionally adds the KMS plugin volume mount to the specified container.
 // It assumes the pod spec does not already contain the KMS volume or mount; no deduplication is performed.
