+++ pkg/operator/encryption/kms/helpers_test.go
@@ -12,108 +12,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestKeyIDFromPluginConfigSecretDataKey(t *testing.T) {
-	tests := []struct {
-		name      string
-		dataKey   string
-		wantKeyID string
-		wantFound bool
-		wantError bool
-	}{
-		{
-			name:      "valid key",
-			dataKey:   "kms-plugin-config-1",
-			wantKeyID: "1",
-			wantFound: true,
-		},
-		{
-			name:      "valid key with large ID",
-			dataKey:   "kms-plugin-config-42",
-			wantKeyID: "42",
-			wantFound: true,
-		},
-		{
-			name:    "encryption-config key",
-			dataKey: "encryption-config",
-		},
-		{
-			name:      "non-integer keyID",
-			dataKey:   "kms-plugin-config-abc",
-			wantError: true,
-		},
-		{
-			name:    "missing keyID",
-			dataKey: "kms-plugin-config-",
-		},
-		{
-			name:    "empty string",
-			dataKey: "",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			keyID, found, err := KeyIDFromPluginConfigSecretDataKey(tt.dataKey)
-			if tt.wantError {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.Equal(t, tt.wantFound, found)
-			if found {
-				require.Equal(t, tt.wantKeyID, keyID)
-			}
-		})
-	}
-}
-
-func TestToPluginConfigSecretDataKeyFor(t *testing.T) {
-	tests := []struct {
-		name      string
-		keyID     string
-		wantKey   string
-		wantError bool
-	}{
-		{
-			name:    "valid keyID",
-			keyID:   "1",
-			wantKey: "kms-plugin-config-1",
-		},
-		{
-			name:    "valid large keyID",
-			keyID:   "42",
-			wantKey: "kms-plugin-config-42",
-		},
-		{
-			name:      "non-integer keyID",
-			keyID:     "abc",
-			wantError: true,
-		},
-		{
-			name:      "empty keyID",
-			keyID:     "",
-			wantError: true,
-		},
-		{
-			name:      "negative keyID",
-			keyID:     "-1",
-			wantError: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := ToPluginConfigSecretDataKeyFor(tt.keyID)
-			if tt.wantError {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.Equal(t, tt.wantKey, got)
-		})
-	}
-}
-
 func TestAddKMSPluginVolume(t *testing.T) {
 	directoryOrCreate := corev1.HostPathDirectoryOrCreate
